requirements.txt
cffi==1.11.5
Django==2.0.6
django-debug-toolbar==1.9.1
greenlet==0.4.13
gunicorn==19.8.1
psycopg2==2.7.5
pytz==2018.5
readline==6.2.4.1
sqlparse==0.2.4
whitenoise==3.3.1
